CAEN code revamped and improved!
It is possible to indicate the CAEN model with its associated constants.
Two different types of setups: one for the global config and a individual channel config.
Events can act as their own data type instead of relying on CAEN functions.
Control software is stable and allows to save to file.
Cleaned the code in a few places.